Add separate package list view model
