+++ WooCommerce/Classes/ViewRelated/Orders/Order Details/Shipping Labels/Create Shipping Label Form/Package Details/ShippingLabelPackageDetailsViewModel.swift
@@ -24,29 +24,13 @@ final class ShippingLabelPackageDetailsViewModel: ObservableObject {
     ///
     @Published private var productVariations: [ProductVariation] = []
 
-    /// The packages  response fetched from API
-    ///
-    @Published private(set) var packagesResponse: ShippingLabelPackagesResponse?
-
-    var dimensionUnit: String {
-        return packagesResponse?.storeOptions.dimensionUnit ?? ""
-    }
-    var customPackages: [ShippingLabelCustomPackage] {
-        return packagesResponse?.customPackages ?? []
-    }
-    var predefinedOptions: [ShippingLabelPredefinedOption] {
-        return packagesResponse?.predefinedOptions ?? []
-    }
-
-    /// Whether there are saved custom or predefined packages to select from.
-    ///
-    var hasCustomOrPredefinedPackages: Bool {
-        return customPackages.isNotEmpty || predefinedOptions.isNotEmpty
-    }
-
     /// The weight unit used in the Store
     ///
     let weightUnit: String?
+
+    /// View model for the package list
+    ///
+    let packageListViewModel: ShippingLabelPackageListViewModel
 
     /// The items rows observed by the main view `ShippingLabelPackageDetails`
     ///
@@ -69,31 +53,22 @@ final class ShippingLabelPackageDetailsViewModel: ObservableObject {
     /// The title of the selected package, if any.
     ///
     var selectedPackageName: String {
-        if let selectedCustomPackage = selectedCustomPackage {
+        if let selectedCustomPackage = packageListViewModel.selectedCustomPackage {
             return selectedCustomPackage.title
         }
-        else if let selectedPredefinedPackage = selectedPredefinedPackage {
+        else if let selectedPredefinedPackage = packageListViewModel.selectedPredefinedPackage {
             return selectedPredefinedPackage.title
         }
         else {
             return Localization.selectPackagePlaceholder
         }
     }
-    @Published private(set) var selectedCustomPackage: ShippingLabelCustomPackage?
-    @Published private(set) var selectedPredefinedPackage: ShippingLabelPredefinedPackage?
+
     @Published var totalWeight: String = ""
 
     /// Whether the user has edited the total package weight. If true, we won't make any automatic changes to the total weight.
     ///
     @Published private var isPackageWeightEdited: Bool = false
-
-    /// Returns if the custom packages header should be shown in Package List
-    ///
-    var showCustomPackagesHeader: Bool {
-        return customPackages.count > 0
-    }
-
-    lazy var addNewPackageViewModel: ShippingLabelAddNewPackageViewModel = ShippingLabelAddNewPackageViewModel(packagesResponse: packagesResponse)
 
     init(order: Order,
          packagesResponse: ShippingLabelPackagesResponse?,
@@ -109,7 +84,7 @@ final class ShippingLabelPackageDetailsViewModel: ObservableObject {
         self.stores = stores
         self.storageManager = storageManager
         self.weightUnit = weightUnit
-        self.packagesResponse = packagesResponse
+        self.packageListViewModel = ShippingLabelPackageListViewModel(packagesResponse: packagesResponse)
         self.selectedPackageID = selectedPackages.first?.packageID // TODO-4599: fix this
 
         configureResultsControllers()
@@ -136,7 +111,7 @@ final class ShippingLabelPackageDetailsViewModel: ObservableObject {
     ///
     private func configureTotalWeights(initialTotalWeight: String?) {
         if let initialTotalWeight = initialTotalWeight {
-            let calculatedWeight = calculateTotalWeight(products: products, productVariations: productVariations, customPackage: selectedCustomPackage)
+            let calculatedWeight = calculateTotalWeight(products: products, productVariations: productVariations, customPackage: packageListViewModel.selectedCustomPackage)
             // Return early if manual input is detected
             if initialTotalWeight != String(calculatedWeight) {
                 isPackageWeightEdited = true
@@ -148,7 +123,7 @@ final class ShippingLabelPackageDetailsViewModel: ObservableObject {
         // This takes into account changes of selected custom package, products and variations.
         // The stream should be completed immediately if manual input of package weight is detected.
         //
-        let calculatedWeight = $selectedCustomPackage.combineLatest($products, $productVariations)
+        let calculatedWeight = packageListViewModel.$selectedCustomPackage.combineLatest($products, $productVariations)
             .map { [weak self] (customPackage, products, variations) -> Double in
                 self?.calculateTotalWeight(products: products, productVariations: variations, customPackage: customPackage) ?? 0
             }
@@ -277,61 +252,20 @@ extension ShippingLabelPackageDetailsViewModel {
 
 // MARK: - Package Selection
 extension ShippingLabelPackageDetailsViewModel {
-    func didSelectPackage(_ id: String) {
-        selectCustomPackage(id)
-        selectPredefinedPackage(id)
-    }
-
-    private func selectCustomPackage(_ id: String) {
-        guard let packagesResponse = packagesResponse else {
-            return
-        }
-
-        for customPackage in packagesResponse.customPackages {
-            if customPackage.title == id {
-                selectedCustomPackage = customPackage
-                selectedPredefinedPackage = nil
-                return
-            }
-        }
-    }
-
-    private func selectPredefinedPackage(_ id: String) {
-        guard let packagesResponse = packagesResponse else {
-            return
-        }
-
-        for option in packagesResponse.predefinedOptions {
-            for predefinedPackage in option.predefinedPackages {
-                if predefinedPackage.id == id {
-                    selectedCustomPackage = nil
-                    selectedPredefinedPackage = predefinedPackage
-                    return
-                }
-            }
-        }
-    }
-
-    /// Writes into the binding variable the final package selection value when confirmed.
-    /// Also sets the total weight for the package, including the selected package weight (if any).
-    ///
-    func confirmPackageSelection() {
-        if let selectedCustomPackage = selectedCustomPackage {
-            selectedPackageID = selectedCustomPackage.title
-        }
-        else if let selectedPredefinedPackage = selectedPredefinedPackage {
-            selectedPackageID = selectedPredefinedPackage.id
-        }
-    }
-
     /// Sets the package passed through the init method, or set the last selected package, if any, as the default selected package
     ///
     func setDefaultPackage() {
         guard let selectedPackageID = selectedPackageID ?? resultsControllers?.accountSettings?.lastSelectedPackageID else {
             return
         }
-        didSelectPackage(selectedPackageID)
-        confirmPackageSelection()
+        packageListViewModel.didSelectPackage(selectedPackageID)
+        packageListViewModel.confirmPackageSelection()
+//        if let selectedCustomPackage = selectedCustomPackage {
+//            selectedPackageID = selectedCustomPackage.title
+//        }
+//        else if let selectedPredefinedPackage = selectedPredefinedPackage {
+//            selectedPackageID = selectedPredefinedPackage.id
+//        }
     }
 }
